+++ paddle/phi/core/distributed/store/tcp_store_libuv.cc
@@ -12,501 +12,467 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include <uv.h>
-
-#include <algorithm>
-#include <cstdio>
-#include <deque>
-#include <exception>
-#include <memory>
-#include <unordered_map>
-#include <unordered_set>
-#include <utility>
-#include <vector>
-
-#include "glog/logging.h"
-
-#include "paddle/common/flags.h"
-#include "paddle/phi/core/distributed/store/tcp_store.h"
-#include "paddle/phi/core/distributed/store/tcp_utils.h"
+#include "paddle/phi/core/distributed/store/tcp_store_libuv.h"
 
 namespace phi::distributed::detail {
 
-auto constexpr MAX_KEY_LEN = 16 * 1024;
-auto constexpr MAX_CONTENT_LEN = 16 * 1024 * 1024;
-auto constexpr MAX_BUFFER_SIZE = size_t(4096);
+// SegmentedDataStream
+void SegmentedDataStream::append(uv_buf_t buf) {
+  if (buf.len == 0) {
+    free(buf.base);
+  } else {
+    capacity += buf.len;
+    _buffers.push_back(buf);
+  }
+}
 
-class LibUVHandle : public std::enable_shared_from_this<LibUVHandle> {
- public:
-  ~LibUVHandle() = default;
-
-  std::shared_ptr<LibUVHandle> ptr() { return shared_from_this(); }
-
-  virtual uv_handle_t* getRawHandle() = 0;
-
-  void close() {
-    if (uv_is_closing(getRawHandle())) {
-      return;
-    }
-    uv_close(getRawHandle(), handleClose);
+bool SegmentedDataStream::readMany(char* dest, size_t size) {
+  if (available() < size) {
+    return false;
   }
 
- protected:
-  void handleAvailable() { uv_handle_set_data(getRawHandle(), this); }
-
-  virtual void onClose() = 0;
-
- private:
-  static void handleClose(uv_handle_t* uv_handle) {
-    auto h = reinterpret_cast<LibUVHandle*>(uv_handle_get_data(uv_handle));
-    h->onClose();
-  }
-};
-
-class LibUVTCPSocket : public LibUVHandle {
- public:
-  explicit LibUVTCPSocket(uv_loop_t* loop) {
-    uv_tcp_init(loop, &client);
-    if (int err = uv_tcp_nodelay(&client, 1)) {
-      VLOG(2)
-          << "The no-delay option is unavailable for the client socket. err: "
-          << err;
-    }
-  }
-  uv_handle_t* getRawHandle() override {
-    return reinterpret_cast<uv_handle_t*>(&client);
-  }
-
-  std::shared_ptr<LibUVTCPSocket> ptr() {
-    return std::static_pointer_cast<LibUVTCPSocket>(shared_from_this());
-  }
-
-  static std::shared_ptr<LibUVTCPSocket> getTCPSocket(uv_stream_t* handle) {
-    auto h = reinterpret_cast<LibUVTCPSocket*>(
-        uv_handle_get_data(reinterpret_cast<uv_handle_t*>(handle)));
-    return h->ptr();
-  }
-
-  virtual void doProcess(const uv_buf_t* buf, size_t nread) {
-    PADDLE_THROW(
-        common::errors::Fatal("Socket subclass does not implement doProcess"));
-  }
-
-  uv_tcp_t client{};
-
- protected:
-  void onClose() override {}
-};
-
-class LibUVTCPServer : public LibUVTCPSocket {
- public:
-  typedef std::function<void(int)> LibUVCallback;
-
-  explicit LibUVTCPServer(uv_loop_t* loop)
-      : LibUVTCPSocket(loop), _on_connect_callback(defaultOnConnect) {}
-
-  void setCallback(LibUVCallback&& callback) {
-    _on_connect_callback = std::move(callback);
-  }
-
-  static std::shared_ptr<LibUVTCPServer> createServer(uv_loop_t* loop,
-                                                      std::uint16_t port,
-                                                      bool useIpv6) {
-    auto res = std::make_shared<LibUVTCPServer>(loop);
-    res->handleAvailable();
-    try {
-      struct sockaddr_storage addr {};
-      int uv_res = 0;
-      if (useIpv6) {
-        uv_res = uv_ip6_addr("::", port, (struct sockaddr_in6*)&addr);
-      } else {
-        uv_res = uv_ip4_addr("0.0.0.0", port, (struct sockaddr_in*)&addr);
+  size_t remaining = size;
+  char* write_base = dest;
+  while (remaining > 0) {
+    auto to_read = std::min(_buffers[_buff_idx].len - _buff_offset, remaining);
+    ::memcpy(write_base, _buffers[_buff_idx].base + _buff_offset, to_read);
+    _buff_offset += to_read;
+    remaining -= to_read;
+    write_base += to_read;
+    if (_buff_offset >= _buffers[_buff_idx].len) {
+      _buff_offset = 0;
+      ++_buff_idx;
+      if (_buff_idx >= _buffers.size() && remaining > 0) {
+        PADDLE_THROW(common::errors::Fatal(
+            paddle::string::Sprintf("Read operation exceeds buffer boundary. "
+                                    "buffer index: %d, available: %d, "
+                                    "remaining: %d",
+                                    _buff_idx,
+                                    _buffers.size(),
+                                    remaining)));
       }
-      PADDLE_ENFORCE_EQ(uv_res,
-                        0,
-                        common::errors::InvalidArgument(paddle::string::Sprintf(
-                            "sockaddr parsing failure. port: %d, useIpv6:%d, "
-                            "code: %d, name: %s, message: %s",
-                            port,
-                            useIpv6,
-                            uv_res,
-                            uv_err_name(uv_res),
-                            uv_strerror(uv_res))));
-
-      uv_res =
-          uv_tcp_bind(res->getRawSocket(), (const struct ::sockaddr*)&addr, 0);
-      PADDLE_ENFORCE_EQ(
-          uv_res,
-          0,
-          common::errors::InvalidArgument(paddle::string::Sprintf(
-              "Bind operation failed for the server socket. port: %d, "
-              "useIpv6: %d, code: %d, name: %s, message: %s",
-              port,
-              useIpv6,
-              uv_res,
-              uv_err_name(uv_res),
-              uv_strerror(uv_res))));
-
-      uv_res = uv_listen(
-          res->getRawStream(), FLAGS_tcp_max_syn_backlog, onNewConnection);
-      PADDLE_ENFORCE_EQ(
-          uv_res,
-          0,
-          common::errors::InvalidArgument(paddle::string::Sprintf(
-              "Server socket unable to listen on local network interfaces. "
-              "port: %d, useIpv6: %d, code: %d, name: %s, message: %s",
-              port,
-              useIpv6,
-              uv_res,
-              uv_err_name(uv_res),
-              uv_strerror(uv_res))));
-      res->setSocketPort();
-    } catch (std::exception& ex) {
-      res->close();
-      throw;
     }
+  }
+  _read_offset += size;
+  return true;
+}
 
-    return res;
+bool SegmentedDataStream::readKey(std::string& str) {
+  uint64_t size = 0;
+  if (!readValue(size)) return false;
+  PADDLE_ENFORCE_LE(size,
+                    phi::distributed::detail::MAX_KEY_LEN,
+                    common::errors::InvalidArgument(paddle::string::Sprintf(
+                        "Key size validation failed. size: %d, max: %d",
+                        size,
+                        phi::distributed::detail::MAX_KEY_LEN)));
+
+  if (available() < size) return false;
+  str.resize(size);
+  return readMany(reinterpret_cast<char*>(str.data()), size);
+}
+
+bool SegmentedDataStream::readContent(std::vector<uint8_t>& data) {
+  uint64_t size = 0;
+  if (!readValue(size)) return false;
+  auto size_in_bytes = size * sizeof(uint8_t);
+  PADDLE_ENFORCE_LE(size_in_bytes,
+                    MAX_CONTENT_LEN,
+                    common::errors::InvalidArgument(paddle::string::Sprintf(
+                        "Content size validation failed. size: %d, max: %d",
+                        size_in_bytes,
+                        MAX_CONTENT_LEN)));
+
+  if (available() < size_in_bytes) return false;
+  data.resize(size);
+  return readMany(reinterpret_cast<char*>(data.data()), size_in_bytes);
+}
+
+size_t SegmentedDataStream::available() { return capacity - _read_offset; }
+
+void SegmentedDataStream::commit() {
+  if (_buff_idx >= _buffers.size() || _buff_offset >= _buffers[_buff_idx].len) {
+    _buff_offset = 0;
+    if (_buff_idx < _buffers.size()) ++_buff_idx;
   }
 
-  std::uint16_t port() const { return _port; }
+  for (size_t i = 0; i < _buff_idx; ++i) {
+    free(_buffers[0].base);
+    capacity -= _buffers[0].len;
+    _buffers.pop_front();
+  }
+  _buff_idx = 0;
+  _read_offset = _buff_offset_commit = _buff_offset;
+}
 
-  void accept(const std::shared_ptr<LibUVTCPSocket>& socket) {
-    int res = uv_accept(getRawStream(),
-                        reinterpret_cast<uv_stream_t*>(socket->getRawHandle()));
+void SegmentedDataStream::reset() {
+  _buff_idx = 0;
+  _read_offset = _buff_offset = _buff_offset_commit;
+}
+
+//    LibUVTCPSocket
+LibUVTCPSocket::LibUVTCPSocket(uv_loop_t* loop) {
+  uv_tcp_init(loop, &client);
+  if (int err = uv_tcp_nodelay(&client, 1)) {
+    VLOG(2) << "The no-delay option is unavailable for the client socket. err: "
+            << err;
+  }
+}
+
+uv_handle_t* LibUVTCPSocket::getRawHandle() override {
+  return reinterpret_cast<uv_handle_t*>(&client);
+}
+
+static std::shared_ptr<LibUVTCPSocket> LibUVTCPSocket::getTCPSocket(
+    uv_stream_t* handle) {
+  auto h = reinterpret_cast<LibUVTCPSocket*>(
+      uv_handle_get_data(reinterpret_cast<uv_handle_t*>(handle)));
+  return h->ptr();
+}
+
+//   LibUVTCPServer
+void LibUVTCPServer::setCallback(LibUVCallback&& callback) {
+  _on_connect_callback = std::move(callback);
+}
+
+static std::shared_ptr<LibUVTCPServer> LibUVTCPServer::createServer(
+    uv_loop_t* loop, std::uint16_t port, bool useIpv6) {
+  auto res = std::make_shared<LibUVTCPServer>(loop);
+  res->handleAvailable();
+  try {
+    struct sockaddr_storage addr {};
+    int uv_res = 0;
+    if (useIpv6) {
+      uv_res = uv_ip6_addr("::", port, (struct sockaddr_in6*)&addr);
+    } else {
+      uv_res = uv_ip4_addr("0.0.0.0", port, (struct sockaddr_in*)&addr);
+    }
+    PADDLE_ENFORCE_EQ(uv_res,
+                      0,
+                      common::errors::InvalidArgument(paddle::string::Sprintf(
+                          "sockaddr parsing failure. port: %d, useIpv6:%d, "
+                          "code: %d, name: %s, message: %s",
+                          port,
+                          useIpv6,
+                          uv_res,
+                          uv_err_name(uv_res),
+                          uv_strerror(uv_res))));
+
+    uv_res =
+        uv_tcp_bind(res->getRawSocket(), (const struct ::sockaddr*)&addr, 0);
     PADDLE_ENFORCE_EQ(
-        res,
+        uv_res,
         0,
         common::errors::InvalidArgument(paddle::string::Sprintf(
-            "Socket accept operation failed. code: %d, name: %s, message: %s",
-            res,
-            uv_err_name(res),
-            uv_strerror(res))));
+            "Bind operation failed for the server socket. port: %d, "
+            "useIpv6: %d, code: %d, name: %s, message: %s",
+            port,
+            useIpv6,
+            uv_res,
+            uv_err_name(uv_res),
+            uv_strerror(uv_res))));
+
+    uv_res = uv_listen(
+        res->getRawStream(), FLAGS_tcp_max_syn_backlog, onNewConnection);
+    PADDLE_ENFORCE_EQ(
+        uv_res,
+        0,
+        common::errors::InvalidArgument(paddle::string::Sprintf(
+            "Server socket unable to listen on local network interfaces. "
+            "port: %d, useIpv6: %d, code: %d, name: %s, message: %s",
+            port,
+            useIpv6,
+            uv_res,
+            uv_err_name(uv_res),
+            uv_strerror(uv_res))));
+    res->setSocketPort();
+  } catch (std::exception& ex) {
+    res->close();
+    throw;
   }
 
- protected:
-  uv_tcp_t* getRawSocket() { return &client; }
-  uv_stream_t* getRawStream() {
-    return reinterpret_cast<uv_stream_t*>(&client);
+  return res;
+}
+
+void LibUVTCPServer::accept(const std::shared_ptr<LibUVTCPSocket>& socket) {
+  int res = uv_accept(getRawStream(),
+                      reinterpret_cast<uv_stream_t*>(socket->getRawHandle()));
+  PADDLE_ENFORCE_EQ(
+      res,
+      0,
+      common::errors::InvalidArgument(paddle::string::Sprintf(
+          "Socket accept operation failed. code: %d, name: %s, message: %s",
+          res,
+          uv_err_name(res),
+          uv_strerror(res))));
+}
+
+void LibUVTCPServer::setSocketPort() {
+  sockaddr_storage addr_s{};
+
+  int addr_len = sizeof(addr_s);
+
+  if (uv_tcp_getsockname(reinterpret_cast<uv_tcp_t*>(getRawStream()),
+                         reinterpret_cast<::sockaddr*>(&addr_s),
+                         &addr_len) != 0) {
+    throw std::runtime_error(
+        "uv_tcp_getsockname failed, the port number cannot be retrieved.");
   }
-
- private:
-  LibUVCallback _on_connect_callback;
-  std::uint16_t _port{};
-
-  void setSocketPort() {
-    sockaddr_storage addr_s{};
-
-    int addr_len = sizeof(addr_s);
-
-    if (uv_tcp_getsockname(reinterpret_cast<uv_tcp_t*>(getRawStream()),
-                           reinterpret_cast<::sockaddr*>(&addr_s),
-                           &addr_len) != 0) {
-      throw std::runtime_error(
-          "uv_tcp_getsockname failed, the port number cannot be retrieved.");
-    }
-    if (addr_s.ss_family == AF_INET) {
-      _port = ntohs(reinterpret_cast<sockaddr_in*>(&addr_s)->sin_port);
-    } else {
-      _port = ntohs(reinterpret_cast<sockaddr_in6*>(&addr_s)->sin6_port);
-    }
+  if (addr_s.ss_family == AF_INET) {
+    _port = ntohs(reinterpret_cast<sockaddr_in*>(&addr_s)->sin_port);
+  } else {
+    _port = ntohs(reinterpret_cast<sockaddr_in6*>(&addr_s)->sin6_port);
   }
+}
 
-  static void defaultOnConnect(int status) {
-    PADDLE_THROW(common::errors::Fatal(
-        "Socket accepted, but onConnect callback is undefined"));
+// WriteUVContent
+WriteUVContent::WriteUVContent(std::vector<uint8_t>&& in_data,
+                               std::shared_ptr<LibUVHandle> handle)
+    : data(std::move(in_data)), handle(std::move(handle)) {
+  uv_req_set_data(reinterpret_cast<uv_req_t*>(&req), new RequestData());
+}
+
+static void WriteUVContent::writeDone(uv_write_t* req, int status) {
+  auto data_ptr = static_cast<RequestData*>(
+      uv_req_get_data(reinterpret_cast<uv_req_t*>(req)));
+  if (!data_ptr) return;
+
+  auto self = std::move(data_ptr->strong_self);
+  delete data_ptr;
+  uv_req_set_data(reinterpret_cast<uv_req_t*>(req), nullptr);
+
+  if (self && status) {
+    VLOG(2) << "Write to client failed. code:" << status
+            << " desc:" << uv_strerror(status)
+            << " name:" << uv_err_name(status);
+    self->handle->close();
   }
+}
 
-  static void onNewConnection(uv_stream_t* server, int status) {
-    auto h = reinterpret_cast<LibUVTCPServer*>(
-        uv_handle_get_data(reinterpret_cast<uv_handle_t*>(server)));
-    h->_on_connect_callback(status);
+WriteUVContent::~WriteUVContent() {
+  // safely clean up pending request data
+  if (auto data = static_cast<RequestData*>(
+          uv_req_get_data(reinterpret_cast<uv_req_t*>(&req)))) {
+    delete data;
+    uv_req_set_data(reinterpret_cast<uv_req_t*>(&req), nullptr);
   }
-};
+}
 
-class LibUVMasterDaemon : public DaemonThread {
- public:
-  explicit LibUVMasterDaemon(int port);
-  // Disable copy constructor
-  LibUVMasterDaemon(const LibUVMasterDaemon& other) = delete;
-  // Disable move constructor
-  LibUVMasterDaemon(LibUVMasterDaemon&& other) = delete;
-  // Disable copy assignment operator
-  LibUVMasterDaemon& operator=(const LibUVMasterDaemon& other) = delete;
-  // Disable move assignment operator
-  LibUVMasterDaemon& operator=(LibUVMasterDaemon&& other) = delete;
+void WriteUVContent::send() {
+  if (data.empty()) return;
 
-  ~LibUVMasterDaemon() override;
+  buf = uv_buf_init(reinterpret_cast<char*>(data.data()), data.size());
+  int res = uv_write(&req,
+                     reinterpret_cast<uv_stream_t*>(handle->getRawHandle()),
+                     &buf,
+                     1,
+                     writeDone);
 
-  void init(const std::uint16_t& port);
-  // operator for key
-  void set(const std::string& key, const std::vector<uint8_t>& value);
-  const std::vector<uint8_t>& get(const std::string& key);
-  int64_t add(const std::string& key, int64_t addVal);
-  bool waitKey(const std::string& key,
-               const std::shared_ptr<LibUVHandle>& client);
-  bool checkKeys(const std::vector<std::string>& keys);
-  int64_t size();
-  void append(const std::string& key, const std::vector<uint8_t>& value);
-
-  // client
-  void addClient(const std::shared_ptr<LibUVHandle>& client);
-  void removeClient(const std::shared_ptr<LibUVHandle>& client);
-  void clearWaitState(const std::shared_ptr<LibUVHandle>& client);
-
- protected:
-  void run() override;
-  void stop() override;
-
- private:
-  uv_loop_t loop_{};
-  uv_async_t _exit_handle{};
-  // tcp server
-  std::shared_ptr<LibUVTCPServer> _tcp_server;
-  // tcp store
-  std::unordered_map<std::string, std::vector<uint8_t>> _tcp_store;
-  // the list of LibUVClient waiting on the key
-  std::unordered_map<std::string, std::vector<std::shared_ptr<LibUVHandle>>>
-      _waiting_sockets;
-  // number of keys awaited
-  std::unordered_map<std::shared_ptr<LibUVHandle>, size_t> _awaited_keys;
-  std::unordered_set<std::shared_ptr<LibUVHandle>> _clients;
-
-  int port_;
-
-  static LibUVMasterDaemon& UVMasterDaemon(uv_handle_t* stream) {
-    return *reinterpret_cast<LibUVMasterDaemon*>(uv_handle_get_data(stream));
-  }
-
-  static void on_new_connection(uv_stream_t* server, int status) {
-    UVMasterDaemon(reinterpret_cast<uv_handle_t*>(server)).onConnect(status);
-  }
-
-  static void on_exit_request(uv_async_t* handle) {
-    UVMasterDaemon(reinterpret_cast<uv_handle_t*>(handle)).onExitRequest();
-  }
-
-  void onConnect(int status);
-  void onExitRequest();
-  void notifyWaitingClients(const std::string& key);
-};
-
-class WriteUVContent : public std::enable_shared_from_this<WriteUVContent> {
-  std::shared_ptr<WriteUVContent> ptr() { return shared_from_this(); }
-
-  static void writeDone(uv_write_t* req, int status) {
+  if (res) {
+    VLOG(2) << "Write failed. code:" << res << "desc:" << uv_strerror(res)
+            << "name:" << uv_err_name(res);
+    handle->close();
+  } else {
+    /* This object was successfully registered with the event loop, so keep it
+     * alive until it's unregistered. */
     auto data_ptr = static_cast<RequestData*>(
-        uv_req_get_data(reinterpret_cast<uv_req_t*>(req)));
-    if (!data_ptr) return;
-
-    auto self = std::move(data_ptr->strong_self);
-    delete data_ptr;
-    uv_req_set_data(reinterpret_cast<uv_req_t*>(req), nullptr);
-
-    if (self && status) {
-      VLOG(2) << "Write to client failed. code:" << status
-              << " desc:" << uv_strerror(status)
-              << " name:" << uv_err_name(status);
-      self->handle->close();
+        uv_req_get_data(reinterpret_cast<uv_req_t*>(&req)));
+    if (data_ptr) {
+      data_ptr->strong_self = shared_from_this();
     }
   }
+}
 
-  struct RequestData {
-    std::shared_ptr<WriteUVContent> strong_self;
-  };
+// LibUVClient
+static void LibUVClient::allocBuffer(uv_handle_t* handle,
+                                     size_t buf_size,
+                                     uv_buf_t* buf) {
+  buf_size = std::min(buf_size, MAX_BUFFER_SIZE);
+  buf->base = reinterpret_cast<char*>(malloc(buf_size));
+  buf->len = buf_size;
+}
 
-  std::vector<uint8_t> data;
-  uv_write_t req = {};
-  uv_buf_t buf = {};
-  std::shared_ptr<LibUVHandle> handle;
+static void LibUVClient::readCallback(uv_stream_t* client,
+                                      ssize_t nread,
+                                      const uv_buf_t* buf) {
+  auto uv_socket = LibUVTCPSocket::getTCPSocket(client);
 
- public:
-  WriteUVContent(std::vector<uint8_t>&& in_data,
-                 std::shared_ptr<LibUVHandle> handle)
-      : data(std::move(in_data)), handle(std::move(handle)) {
-    uv_req_set_data(reinterpret_cast<uv_req_t*>(&req), new RequestData());
-  }
-
-  ~WriteUVContent() {
-    // safely clean up pending request data
-    if (auto data = static_cast<RequestData*>(
-            uv_req_get_data(reinterpret_cast<uv_req_t*>(&req)))) {
-      delete data;
-      uv_req_set_data(reinterpret_cast<uv_req_t*>(&req), nullptr);
+  if (nread > 0) {
+    try {
+      uv_socket->doProcess(buf, nread);
+      return;
+    } catch (std::exception& ex) {
+      VLOG(2) << "Failed to process incoming client message: " << ex.what();
+      uv_socket->close();
     }
+  } else if (nread == UV_EOF) {
+    // EOF
+    VLOG(5) << "Remote peer closed the connection.";
+    uv_socket->close();
+  } else if (nread < 0) {
+    // error and EOF
+    VLOG(5) << "Read callback handler exception. code:" << nread
+            << " desc:" << uv_strerror(nread) << " name:" << uv_err_name(nread);
+    uv_socket->close();
   }
+  free(buf->base);
+}
 
-  void send() {
-    if (data.empty()) return;
+void LibUVClient::doProcess(const uv_buf_t* buf, size_t nread) override {
+  auto tmp = *buf;
+  tmp.len = nread;
+  stream.append(tmp);
 
-    buf = uv_buf_init(reinterpret_cast<char*>(data.data()), data.size());
-    int res = uv_write(&req,
-                       reinterpret_cast<uv_stream_t*>(handle->getRawHandle()),
-                       &buf,
-                       1,
-                       writeDone);
+  VLOG(5) << "process: " << std::string(buf->base, nread)
+          << ", nread: " << nread;
+  while (true) {
+    stream.reset();
+    uint32_t command = -1;
+    if (!stream.readValue(command)) break;
 
-    if (res) {
-      VLOG(2) << "Write failed. code:" << res << "desc:" << uv_strerror(res)
-              << "name:" << uv_err_name(res);
-      handle->close();
-    } else {
-      /* This object was successfully registered with the event loop, so keep it
-       * alive until it's unregistered. */
-      auto data_ptr = static_cast<RequestData*>(
-          uv_req_get_data(reinterpret_cast<uv_req_t*>(&req)));
-      if (data_ptr) {
-        data_ptr->strong_self = shared_from_this();
-      }
-    }
-  }
-};
-
-class UVWriter {
-  std::vector<uint8_t> data;
-  std::shared_ptr<LibUVHandle> handle;
-
-  void* operator new(size_t);
-
- public:
-  explicit UVWriter(std::shared_ptr<LibUVHandle> handle)
-      : handle(std::move(handle)) {}
-
-  template <typename T>
-  void writeValue(T val) {
-    uint8_t* val_ptr = reinterpret_cast<uint8_t*>(&val);
-    data.insert(data.end(), val_ptr, val_ptr + sizeof(T));
-  }
-
-  void writeVector(const std::vector<uint8_t>& val) {
-    writeValue<uint64_t>(val.size());
-    data.insert(data.end(), val.begin(), val.end());
-  }
-
-  void writeString(const std::string& val) {
-    writeValue<uint64_t>(val.size());
-    data.insert(data.end(), val.data(), val.data() + val.size());
-  }
-
-  void send() {
-    auto wd = std::make_shared<WriteUVContent>(std::move(data), handle);
-    wd->send();
-  }
-};
-
-class SegmentedDataStream {
-  std::deque<uv_buf_t> _buffers;
-  size_t _buff_idx{0};
-  size_t _buff_offset{0};
-  size_t capacity{0};
-  size_t _buff_offset_commit{0};
-  size_t _read_offset{0};
-
- public:
-  SegmentedDataStream() = default;
-
-  void append(uv_buf_t buf) {
-    if (buf.len == 0) {
-      free(buf.base);
-    } else {
-      capacity += buf.len;
-      _buffers.push_back(buf);
-    }
-  }
-
-  bool readMany(char* dest, size_t size) {
-    if (available() < size) {
-      return false;
+    VLOG(5) << "Client parse command" << command;
+    switch ((Command)command) {
+      case Command::ADD:
+        if (!doAddCommand()) return;
+        break;
+      case Command::GET:
+        if (!doGetCommand()) return;
+        break;
+      case Command::CHECK:
+        if (!doCheckCommand()) return;
+        break;
+      case Command::SET:
+        if (!doSetCommand()) return;
+        break;
+      case Command::WAIT:
+        if (!doWaitCommand()) return;
+        break;
+      default:
+        VLOG(4) << "invalid command from Client, command: "
+                << static_cast<int>(command);
+        close();
+        return;
     }
 
-    size_t remaining = size;
-    char* write_base = dest;
-    while (remaining > 0) {
-      auto to_read =
-          std::min(_buffers[_buff_idx].len - _buff_offset, remaining);
-      ::memcpy(write_base, _buffers[_buff_idx].base + _buff_offset, to_read);
-      _buff_offset += to_read;
-      remaining -= to_read;
-      write_base += to_read;
-      if (_buff_offset >= _buffers[_buff_idx].len) {
-        _buff_offset = 0;
-        ++_buff_idx;
-        if (_buff_idx >= _buffers.size() && remaining > 0) {
-          PADDLE_THROW(common::errors::Fatal(
-              paddle::string::Sprintf("Read operation exceeds buffer boundary. "
-                                      "buffer index: %d, available: %d, "
-                                      "remaining: %d",
-                                      _buff_idx,
-                                      _buffers.size(),
-                                      remaining)));
-        }
-      }
-    }
-    _read_offset += size;
-    return true;
+    stream.commit();
+  }
+}
+
+bool LibUVClient::doSetCommand() {
+  std::string key;
+  if (!stream.readKey(key)) return false;
+
+  std::vector<uint8_t> newData;
+  if (!stream.readContent(newData)) return false;
+
+  VLOG(7) << "set key:" << key << " address:" << this->address();
+
+  store->set(key, newData);
+  return true;
+}
+
+bool LibUVClient::doGetCommand() {
+  std::string key;
+  if (!stream.readKey(key)) return false;
+
+  VLOG(7) << "get key: " << key << " address:" << this->address();
+
+  const auto& data = store->get(key);
+  UVWriter sw(ptr());
+  sw.writeVector(data);
+  sw.send();
+  return true;
+}
+
+bool LibUVClient::doAddCommand() {
+  std::string key;
+  if (!stream.readKey(key)) return false;
+
+  int64_t addVal = 0;
+  if (!stream.readValue(addVal)) return false;
+
+  addVal = store->add(key, addVal);
+  VLOG(7) << "add key:" << key << " val: " << addVal
+          << " address:" << this->address();
+
+  UVWriter sw(ptr());
+  sw.writeValue(addVal);
+  sw.send();
+
+  return true;
+}
+
+bool LibUVClient::doCheckCommand() {
+  std::string key;
+  if (!stream.readKey(key)) return false;
+
+  VLOG(7) << "check key:" << key << " address:" << this->address();
+
+  // check keys
+  std::vector<std::string> keys = {key};
+  UVWriter sw(ptr());
+  if (store->checkKeys(keys)) {
+    sw.writeValue(ReplyType::READY);
+  } else {
+    sw.writeValue(ReplyType::NOT_READY);
+  }
+  sw.send();
+  return true;
+}
+
+bool LibUVClient::doWaitCommand() {
+  std::string key;
+  if (!stream.readKey(key)) return false;
+
+  VLOG(7) << "wait key:  " << key << " address:" << this->address();
+  if (store->waitKey(key, ptr())) {
+    UVWriter sw(ptr());
+    sw.writeValue(ReplyType::STOP_WAIT);
+    sw.send();
+    VLOG(7) << "wait send:  " << key;
+  }
+  return true;
+}
+
+void LibUVClient::onClose() override { store->removeClient(ptr()); }
+
+void LibUVClient::readStart() {
+  struct ::sockaddr_storage addr {};
+  int addrLen{sizeof(struct ::sockaddr_storage)};
+
+  if (int err = uv_tcp_getpeername(
+          &client, reinterpret_cast<struct ::sockaddr*>(&addr), &addrLen)) {
+    VLOG(2) << "Client remote endpoint resolution failed. err="
+            << uv_strerror(err);
+  } else {
+    _address =
+        formatSockAddr(reinterpret_cast<struct ::sockaddr*>(&addr), addrLen);
   }
 
-  template <typename T>
-  bool readValue(T& value) {  // NOLINT(runtime/references)
-    return readMany(reinterpret_cast<char*>(&value), sizeof(T));
+  int res = uv_read_start(
+      reinterpret_cast<uv_stream_t*>(&client), allocBuffer, readCallback);
+  if (res) {
+    VLOG(2) << "Read callback initialization failure. client:"
+            << reinterpret_cast<void*>(this) << " code:" << res
+            << " desc:" << uv_strerror(res) << " name:" << uv_err_name(res);
+    close();
   }
+}
 
-  bool readKey(std::string& str) {  // NOLINT(runtime/references)
-    uint64_t size = 0;
-    if (!readValue(size)) return false;
-    PADDLE_ENFORCE_LE(size,
-                      phi::distributed::detail::MAX_KEY_LEN,
-                      common::errors::InvalidArgument(paddle::string::Sprintf(
-                          "Key size validation failed. size: %d, max: %d",
-                          size,
-                          phi::distributed::detail::MAX_KEY_LEN)));
+static std::shared_ptr<LibUVClient> LibUVClient::make(
+    uv_loop_t* loop, LibUVMasterDaemon* store) {
+  auto res = std::make_shared<LibUVClient>(loop, store);
+  res->handleAvailable();
+  return res;
+}
 
-    if (available() < size) return false;
-    str.resize(size);
-    return readMany(reinterpret_cast<char*>(str.data()), size);
-  }
+std::shared_ptr<LibUVClient> LibUVClient::ptr() {
+  return std::static_pointer_cast<LibUVClient>(shared_from_this());
+}
 
-  bool readContent(std::vector<uint8_t>& data) {  // NOLINT(runtime/references)
-    uint64_t size = 0;
-    if (!readValue(size)) return false;
-    auto size_in_bytes = size * sizeof(uint8_t);
-    PADDLE_ENFORCE_LE(size_in_bytes,
-                      MAX_CONTENT_LEN,
-                      common::errors::InvalidArgument(paddle::string::Sprintf(
-                          "Content size validation failed. size: %d, max: %d",
-                          size_in_bytes,
-                          MAX_CONTENT_LEN)));
-
-    if (available() < size_in_bytes) return false;
-    data.resize(size);
-    return readMany(reinterpret_cast<char*>(data.data()), size_in_bytes);
-  }
-
-  size_t available() { return capacity - _read_offset; }
-
-  void commit() {
-    if (_buff_idx >= _buffers.size() ||
-        _buff_offset >= _buffers[_buff_idx].len) {
-      _buff_offset = 0;
-      if (_buff_idx < _buffers.size()) ++_buff_idx;
-    }
-
-    for (size_t i = 0; i < _buff_idx; ++i) {
-      free(_buffers[0].base);
-      capacity -= _buffers[0].len;
-      _buffers.pop_front();
-    }
-    _buff_idx = 0;
-    _read_offset = _buff_offset_commit = _buff_offset;
-  }
-
-  void reset() {
-    _buff_idx = 0;
-    _read_offset = _buff_offset = _buff_offset_commit;
-  }
-};
-
-std::string formatSockAddr(const struct ::sockaddr* addr, socklen_t len) {
+//  LibUVMasterDaemon
+PADDLE_API std::string formatSockAddr(const struct ::sockaddr* addr,
+                                      socklen_t len) {
   char host[NI_MAXHOST], port[NI_MAXSERV];  // NOLINT
 
   int flags = NI_NUMERICSERV;
@@ -532,205 +498,6 @@ std::string formatSockAddr(const struct ::sockaddr* addr, socklen_t len) {
       return paddle::string::Sprintf("[%s]:%s", host, port);
   }
 }
-
-class LibUVClient : public LibUVTCPSocket {
-  SegmentedDataStream stream;
-  LibUVMasterDaemon* store;
-  std::string _address{"null"};
-
-  const std::string& address() const { return _address; }
-
-  static void allocBuffer(uv_handle_t* handle, size_t buf_size, uv_buf_t* buf) {
-    buf_size = std::min(buf_size, MAX_BUFFER_SIZE);
-    buf->base = reinterpret_cast<char*>(malloc(buf_size));
-    buf->len = buf_size;
-  }
-
-  static void readCallback(uv_stream_t* client,
-                           ssize_t nread,
-                           const uv_buf_t* buf) {
-    auto uv_socket = LibUVTCPSocket::getTCPSocket(client);
-
-    if (nread > 0) {
-      try {
-        uv_socket->doProcess(buf, nread);
-        return;
-      } catch (std::exception& ex) {
-        VLOG(2) << "Failed to process incoming client message: " << ex.what();
-        uv_socket->close();
-      }
-    } else if (nread == UV_EOF) {
-      // EOF
-      VLOG(5) << "Remote peer closed the connection.";
-      uv_socket->close();
-    } else if (nread < 0) {
-      // error and EOF
-      VLOG(5) << "Read callback handler exception. code:" << nread
-              << " desc:" << uv_strerror(nread)
-              << " name:" << uv_err_name(nread);
-      uv_socket->close();
-    }
-    free(buf->base);
-  }
-
- protected:
-  void doProcess(const uv_buf_t* buf, size_t nread) override {
-    auto tmp = *buf;
-    tmp.len = nread;
-    stream.append(tmp);
-
-    VLOG(5) << "process: " << std::string(buf->base, nread)
-            << ", nread: " << nread;
-    while (true) {
-      stream.reset();
-      uint32_t command = -1;
-      if (!stream.readValue(command)) break;
-
-      VLOG(5) << "Client parse command" << command;
-      switch ((Command)command) {
-        case Command::ADD:
-          if (!doAddCommand()) return;
-          break;
-        case Command::GET:
-          if (!doGetCommand()) return;
-          break;
-        case Command::CHECK:
-          if (!doCheckCommand()) return;
-          break;
-        case Command::SET:
-          if (!doSetCommand()) return;
-          break;
-        case Command::WAIT:
-          if (!doWaitCommand()) return;
-          break;
-        default:
-          VLOG(4) << "invalid command from Client, command: "
-                  << static_cast<int>(command);
-          close();
-          return;
-      }
-
-      stream.commit();
-    }
-  }
-
-  bool doSetCommand() {
-    std::string key;
-    if (!stream.readKey(key)) return false;
-
-    std::vector<uint8_t> newData;
-    if (!stream.readContent(newData)) return false;
-
-    VLOG(7) << "set key:" << key << " address:" << this->address();
-
-    store->set(key, newData);
-    return true;
-  }
-
-  bool doGetCommand() {
-    std::string key;
-    if (!stream.readKey(key)) return false;
-
-    VLOG(7) << "get key: " << key << " address:" << this->address();
-
-    const auto& data = store->get(key);
-    UVWriter sw(ptr());
-    sw.writeVector(data);
-    sw.send();
-    return true;
-  }
-
-  bool doAddCommand() {
-    std::string key;
-    if (!stream.readKey(key)) return false;
-
-    int64_t addVal = 0;
-    if (!stream.readValue(addVal)) return false;
-
-    addVal = store->add(key, addVal);
-    VLOG(7) << "add key:" << key << " val: " << addVal
-            << " address:" << this->address();
-
-    UVWriter sw(ptr());
-    sw.writeValue(addVal);
-    sw.send();
-
-    return true;
-  }
-
-  bool doCheckCommand() {
-    std::string key;
-    if (!stream.readKey(key)) return false;
-
-    VLOG(7) << "check key:" << key << " address:" << this->address();
-
-    // check keys
-    std::vector<std::string> keys = {key};
-    UVWriter sw(ptr());
-    if (store->checkKeys(keys)) {
-      sw.writeValue(ReplyType::READY);
-    } else {
-      sw.writeValue(ReplyType::NOT_READY);
-    }
-    sw.send();
-    return true;
-  }
-
-  bool doWaitCommand() {
-    std::string key;
-    if (!stream.readKey(key)) return false;
-
-    VLOG(7) << "wait key:  " << key << " address:" << this->address();
-    if (store->waitKey(key, ptr())) {
-      UVWriter sw(ptr());
-      sw.writeValue(ReplyType::STOP_WAIT);
-      sw.send();
-      VLOG(7) << "wait send:  " << key;
-    }
-    return true;
-  }
-
- public:
-  explicit LibUVClient(uv_loop_t* loop, LibUVMasterDaemon* store)
-      : LibUVTCPSocket(loop), store(store) {}
-
-  void readStart() {
-    struct ::sockaddr_storage addr {};
-    int addrLen{sizeof(struct ::sockaddr_storage)};
-
-    if (int err = uv_tcp_getpeername(
-            &client, reinterpret_cast<struct ::sockaddr*>(&addr), &addrLen)) {
-      VLOG(2) << "Client remote endpoint resolution failed. err="
-              << uv_strerror(err);
-    } else {
-      _address =
-          formatSockAddr(reinterpret_cast<struct ::sockaddr*>(&addr), addrLen);
-    }
-
-    int res = uv_read_start(
-        reinterpret_cast<uv_stream_t*>(&client), allocBuffer, readCallback);
-    if (res) {
-      VLOG(2) << "Read callback initialization failure. client:"
-              << reinterpret_cast<void*>(this) << " code:" << res
-              << " desc:" << uv_strerror(res) << " name:" << uv_err_name(res);
-      close();
-    }
-  }
-
-  static std::shared_ptr<LibUVClient> make(uv_loop_t* loop,
-                                           LibUVMasterDaemon* store) {
-    auto res = std::make_shared<LibUVClient>(loop, store);
-    res->handleAvailable();
-    return res;
-  }
-
-  std::shared_ptr<LibUVClient> ptr() {
-    return std::static_pointer_cast<LibUVClient>(shared_from_this());
-  }
-
- protected:
-  void onClose() override { store->removeClient(ptr()); }
-};
 
 void LibUVMasterDaemon::onConnect(int status) {
   auto client = LibUVClient::make(&loop_, this);
